pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "raft-chat"
version = "1.0.0"
description = "RAG chat application with document processing"
authors = [{name = "Developer", email = "dev@example.com"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "gradio>=4.7.0",
    "requests>=2.31.0",
    "chromadb>=0.4.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "pymupdf>=1.23.0",
    "pandas>=2.0.0",
    "ollama>=0.1.7",
    "openai>=1.10.0,<2.0.0",
    "python-dotenv>=1.0.0",
    "langchain-openai>=0.0.5",
    "sqlalchemy>=2.0.23",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.1",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.7.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "pytest>=7.0.0",
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "gradio.*",
    "langchain.*",
    "chromadb.*",
    "fitz.*",
    "pandas.*",
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py39"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "B904",  # raise exceptions with raise ... from err
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"app/main.py" = ["B008", "C901"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    "venv",
] 